% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contingency_table.R
\name{calculate_contingency_table}
\alias{calculate_contingency_table}
\title{calculate list of contingency table}
\usage{
calculate_contingency_table(data, var1, var2)
}
\arguments{
\item{data}{data frame}

\item{var1}{column name of the data}

\item{var2}{column name of the data}
}
\value{
a list
}
\description{
store and return a list of contingency tables
}
