% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot.R
\name{create_plot}
\alias{create_plot}
\title{create different types of plots}
\usage{
create_plot(data, plot_type, x, y)
}
\arguments{
\item{data}{data frame}

\item{plot_type}{type of the plot you want}

\item{x}{column name of the data}

\item{y}{column name of the data}
}
\value{
a scatter plot or residual plot, or histogram or qq plot
}
\description{
Function to create different types of plots
}
