% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_and_summarize_lm.R
\name{fit_and_summarize_lm}
\alias{fit_and_summarize_lm}
\title{fit and summarize a linear regression model}
\usage{
fit_and_summarize_lm(data, formula)
}
\arguments{
\item{data}{data frame}

\item{formula}{formula of the linear regression}
}
\value{
summary of linear regression model
}
\description{
Function to fit and summarize a linear regression model
}
