% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam_binomial.R
\name{fit_gam_binomial}
\alias{fit_gam_binomial}
\title{fit a generalized additive model}
\usage{
fit_gam_binomial(data, formula)
}
\arguments{
\item{data}{data frame}

\item{formula}{formula of the linear regression}
}
\value{
a list
}
\description{
Function to fit a generalized additive model (GAM) with binomial family
}
