% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_boxplot.R
\name{generate_boxplot}
\alias{generate_boxplot}
\title{Boxplot generation function}
\usage{
generate_boxplot(data, x_var, fill_var, x_label)
}
\arguments{
\item{data}{data frame}

\item{x_var}{column name of the data}

\item{fill_var}{column name of the data}

\item{x_label}{text as character}
}
\value{
a boxplot
}
\description{
Create a boxplot
}
