% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scatter_plot_with_regression.R
\name{generate_scatter_plot_with_regression}
\alias{generate_scatter_plot_with_regression}
\title{generate scatter plot with regression}
\usage{
generate_scatter_plot_with_regression(data, x_var, y_var)
}
\arguments{
\item{data}{data frame}

\item{x_var}{column name of the data}

\item{y_var}{column name of the data}
}
\value{
a scatter plot
}
\description{
scatter plot with linear regression
}
