% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_t_test.R
\name{perform_t_test}
\alias{perform_t_test}
\title{Compute t test}
\usage{
perform_t_test(data, group_var, value_var, group1, group2, alternative)
}
\arguments{
\item{data}{data frame}

\item{group_var}{column name of the data}

\item{value_var}{column name of the data}

\item{group1}{level of a column}

\item{group2}{level of a column}

\item{alternative}{type of hypothesis}
}
\value{
summary of t test
}
\description{
Compute a t test
}
